requirements (6).txt
agno
firecrawl-py==1.9.0
pydantic
streamlit
openai
python-dotenv